+++ src/runtime/defs_openbsd_386.go
@@ -138,8 +138,8 @@ type timespec struct {
 	tv_nsec int32
 }
 
-func (ts *timespec) set_sec(x int32) {
-	ts.tv_sec = int64(x)
+func (ts *timespec) set_sec(x int64) {
+	ts.tv_sec = x
 }
 
 func (ts *timespec) set_nsec(x int32) {
